source/applications/lib/database/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt
    # via database (pyproject.toml)
aerich==0.9.2
    # via database (pyproject.toml)
aiosqlite==0.21.0
    # via tortoise-orm
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via aerich
asyncclick==8.3.0.7
    # via aerich
asyncpg==0.30.0
    # via database (pyproject.toml)
charset-normalizer==3.4.4
    # via core
deprecated==1.3.1
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-semantic-conventions
dictdiffer==0.9.0
    # via aerich
googleapis-common-protos==1.72.0
    # via opentelemetry-exporter-otlp-proto-grpc
grpcio==1.76.0
    # via opentelemetry-exporter-otlp-proto-grpc
idna==3.11
    # via anyio
importlib-metadata==8.6.1
    # via opentelemetry-api
iso8601==2.1.0
    # via tortoise-orm
opentelemetry-api==1.32.1
    # via
    #   core
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp-proto-common==1.32.1
    # via opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-grpc==1.32.1
    # via core
opentelemetry-instrumentation==0.53b1
    # via opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-httpx==0.53b1
    # via core
opentelemetry-proto==1.32.1
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk==1.32.1
    # via
    #   core
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-semantic-conventions==0.53b1
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-sdk
opentelemetry-util-http==0.53b1
    # via opentelemetry-instrumentation-httpx
packaging==25.0
    # via opentelemetry-instrumentation
protobuf==5.29.5
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
pydantic==2.11.10
    # via
    #   database (pyproject.toml)
    #   core
pydantic-core==2.33.2
    # via pydantic
pypika-tortoise==0.6.2
    # via tortoise-orm
pytz==2025.2
    # via tortoise-orm
sniffio==1.3.1
    # via anyio
tortoise-orm==0.25.1
    # via
    #   database (pyproject.toml)
    #   aerich
typing-extensions==4.15.0
    # via
    #   aiosqlite
    #   anyio
    #   grpcio
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
wrapt==1.17.3
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-httpx
zipp==3.23.0
    # via importlib-metadata
